// src/generated/models/patient-registration-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Healix Clinic Management API
 * REST API for clinic management system with EHR support for Maharashtra small to mid-sized clinics.  ## Features - Patient registration and management - Emergency contact management - Insurance details tracking - Medical history records 
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@healix.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { BloodGroup } from './blood-group';
// May contain unused imports in some cases
// @ts-ignore
import type { EmergencyContactRequest } from './emergency-contact-request';
// May contain unused imports in some cases
// @ts-ignore
import type { Gender } from './gender';
// May contain unused imports in some cases
// @ts-ignore
import type { InsuranceRequest } from './insurance-request';
// May contain unused imports in some cases
// @ts-ignore
import type { MedicalHistoryRequest } from './medical-history-request';
// May contain unused imports in some cases
// @ts-ignore
import type { PreferredContactMethod } from './preferred-contact-method';

export interface PatientRegistrationRequest {
    'firstName': string;
    'lastName': string;
    'dateOfBirth': string;
    'gender': Gender;
    'bloodGroup'?: BloodGroup;
    'aadharNumber'?: string;
    'photoUrl'?: string;
    'mobileNumber': string;
    'emailId'?: string;
    'preferredContactMethod'?: PreferredContactMethod;
    'addressLine1'?: string;
    'city': string;
    'district'?: string;
    'state'?: string;
    'pinCode': string;
    'emergencyContacts'?: Array<EmergencyContactRequest>;
    'insurance'?: InsuranceRequest;
    'medicalHistory'?: MedicalHistoryRequest;
}